requirements.txt
torch>=2.2
pandas
numpy
scikit-learn
matplotlib
tqdm
pyyaml
seaborn
wordcloud
